Fix clang warnings and other minor stylistic issues
Some of the warnings indicated faulty code, which is fixed.

+++ source/main.cpp
@@ -82,11 +82,11 @@ int main(int argc, char** argv)
 
     ret = t1 * t2;
 
-    std::printf("%f\n", ret.y);
+    std::printf("%f\n", static_cast<double>(ret.y));
 
     ret = t1 + t2;
 
-    std::printf("%f\n", ret.y);
+    std::printf("%f\n", static_cast<double>(ret.y));
 
 //    sh3_config config;
 //    sh3_arc arc;
@@ -96,7 +96,6 @@ int main(int argc, char** argv)
 //    arc.LoadFile("data/eff_tex/fire00_tr.pic", data);
 //
 //    sh3_window* window = new sh3_window(640, 480, "SILENT HILL 3");
-
 
     exit(exit_code::SUCCESS);
 }
